10.1 - updated some moments according to advices, added additional check before removal book from db

+++ src/main/java/yevhen/bookstore/service/book/BookServiceImpl.java
@@ -12,11 +12,13 @@ import yevhen.bookstore.dto.book.BookDto;
 import yevhen.bookstore.dto.book.BookDtoWithoutCategoryIds;
 import yevhen.bookstore.dto.book.BookSearchParametersDto;
 import yevhen.bookstore.dto.book.CreateBookRequestDto;
+import yevhen.bookstore.exception.DataProcessingException;
 import yevhen.bookstore.exception.EntityNotFoundException;
 import yevhen.bookstore.mapper.BookMapper;
 import yevhen.bookstore.model.Book;
 import yevhen.bookstore.model.Category;
 import yevhen.bookstore.repository.BookRepository;
+import yevhen.bookstore.repository.CartItemRepository;
 import yevhen.bookstore.repository.specification.BookSpecificationBuilder;
 
 @RequiredArgsConstructor
@@ -25,6 +27,7 @@ public class BookServiceImpl implements BookService {
     private final BookRepository bookRepository;
     private final BookMapper bookMapper;
     private final BookSpecificationBuilder bookSpecificationBuilder;
+    private final CartItemRepository cartItemRepository;
 
     @Override
     public BookDto save(CreateBookRequestDto requestDto) {
@@ -59,6 +62,12 @@ public class BookServiceImpl implements BookService {
 
     @Override
     public void deleteBookById(Long id) {
+        if (cartItemRepository.findAllByBookId(id).stream()
+                        .findAny()
+                        .isPresent()) {
+            throw new DataProcessingException("Removal is forbidden."
+                    + "This book is in some shopping cart");
+        }
         bookRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException("Can't find book by id:" + id));
         bookRepository.deleteById(id);
